+++ Cask/remindian.rb
@@ -1,6 +1,6 @@
 cask "remindian" do
-  version "5.7.0"
-  sha256 "2fb7ce38856801bc914ff6d54b13514442d4966d671d14e79c10c5c4f73949ba"
+  version "5.7.1"
+  sha256 "f1023ac75d45a595d74ccc96cda88e247e1740fde91c3abdce8922eabd07ef96"
 
   url "https://github.com/Santofer/Remindian/releases/download/v#{version}/Remindian-v#{version}.dmg"
   name "Remindian"